host and uart edits

+++ dist-fs/comms/comms.h
@@ -50,7 +50,7 @@ typedef struct comm_driver_t {
    * @return bytes read, < 0 on failure
    */
   int (*read)(comm_context_t *ctx,
-              uint32_t *rx,
+              uint8_t *rx,
               uint16_t rx_sz,
               uint16_t timeout_ms);
   /**
@@ -62,7 +62,7 @@ typedef struct comm_driver_t {
    *
    * @return bytes written, < 0 on failure
    */
-  int (*write_one)(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms);
+  int (*write_one)(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
   /**
    * @brief write a buffer of bytes to the comms stream
    *
@@ -74,7 +74,7 @@ typedef struct comm_driver_t {
    * @return bytes written, < 0 on failure
    */
   int (*write)(comm_context_t *ctx,
-               uint32_t *tx,
+               uint8_t *tx,
                uint16_t tx_sz,
                uint16_t timeout_ms);
   /**
@@ -84,7 +84,7 @@ typedef struct comm_driver_t {
    * @param[in]     opcode  ioctl operation code
    * @param[in/out] data    return data
    */
-  int (*ioctl)(comm_context_t *ctx, uint32_t opcode, void *data);
+  int (*ioctl)(comm_context_t *ctx, uint8_t opcode, void *data);
 } comm_driver_t;
 
 /**
