+++ dist-fs/comms/i2c.c
@@ -5,15 +5,15 @@
 static int i2c_init(comm_context_t *ctx);
 static int i2c_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int i2c_read(comm_context_t *ctx,
-                    uint32_t *rx,
+                    uint8_t *rx,
                     uint16_t rx_sz,
                     uint16_t timeout_ms);
-static int i2c_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms);
+static int i2c_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int i2c_write(comm_context_t *ctx,
-                     uint32_t *tx,
+                     uint8_t *tx,
                      uint16_t tx_size,
                      uint16_t timeout_ms);
-static int i2c_ioctl(comm_context_t *ctx, uint32_t opcode, void *data);
+static int i2c_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
 
 comm_driver_t i2c_ops = {
@@ -41,7 +41,7 @@ static int i2c_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 }
 
 static int i2c_read(comm_context_t *ctx,
-                    uint32_t *rx,
+                    uint8_t *rx,
                     uint16_t rx_sz,
                     uint16_t timeout_ms) {
   (void)ctx;
@@ -55,7 +55,7 @@ static int i2c_read(comm_context_t *ctx,
 
 
 static int
-i2c_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
+i2c_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
   (void)timeout_ms;
@@ -65,7 +65,7 @@ i2c_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
 }
 
 static int i2c_write(comm_context_t *ctx,
-                     uint32_t *tx,
+                     uint8_t *tx,
                      uint16_t tx_size,
                      uint16_t timeout_ms) {
   (void)ctx;
@@ -77,7 +77,7 @@ static int i2c_write(comm_context_t *ctx,
   return bytes_written;
 }
 
-static int i2c_ioctl(comm_context_t *ctx, uint32_t opcode, void *data) {
+static int i2c_ioctl(comm_context_t *ctx, uint8_t opcode, void *data) {
   (void)ctx;
   (void)opcode;
   (void)data;
