+++ dist-fs/comms/network.c
@@ -5,16 +5,16 @@
 static int network_init(comm_context_t *ctx);
 static int network_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int network_read(comm_context_t *ctx,
-                        uint32_t *rx,
+                        uint8_t *rx,
                         uint16_t rx_sz,
                         uint16_t timeout_ms);
 static int
-network_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms);
+network_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int network_write(comm_context_t *ctx,
-                         uint32_t *tx,
+                         uint8_t *tx,
                          uint16_t tx_size,
                          uint16_t timeout_ms);
-static int network_ioctl(comm_context_t *ctx, uint32_t opcode, void *data);
+static int network_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
 comm_driver_t network_ops = {
   .init      = network_init,
@@ -41,7 +41,7 @@ static int network_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 }
 
 static int network_read(comm_context_t *ctx,
-                        uint32_t *rx,
+                        uint8_t *rx,
                         uint16_t rx_sz,
                         uint16_t timeout_ms) {
   (void)ctx;
@@ -55,7 +55,7 @@ static int network_read(comm_context_t *ctx,
 
 
 static int
-network_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
+network_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
   (void)timeout_ms;
@@ -65,7 +65,7 @@ network_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
 }
 
 static int network_write(comm_context_t *ctx,
-                         uint32_t *tx,
+                         uint8_t *tx,
                          uint16_t tx_size,
                          uint16_t timeout_ms) {
   (void)ctx;
@@ -77,7 +77,7 @@ static int network_write(comm_context_t *ctx,
   return bytes_written;
 }
 
-static int network_ioctl(comm_context_t *ctx, uint32_t opcode, void *data) {
+static int network_ioctl(comm_context_t *ctx, uint8_t opcode, void *data) {
   (void)ctx;
   (void)opcode;
   (void)data;
