+++ dist-fs/comms/spi.c
@@ -5,15 +5,15 @@
 static int spi_init(comm_context_t *ctx);
 static int spi_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int spi_read(comm_context_t *ctx,
-                    uint32_t *rx,
+                    uint8_t *rx,
                     uint16_t rx_sz,
                     uint16_t timeout_ms);
-static int spi_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms);
+static int spi_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int spi_write(comm_context_t *ctx,
-                     uint32_t *tx,
+                     uint8_t *tx,
                      uint16_t tx_size,
                      uint16_t timeout_ms);
-static int spi_ioctl(comm_context_t *ctx, uint32_t opcode, void *data);
+static int spi_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
 comm_driver_t spi_ops = {
   .init      = spi_init,
@@ -40,7 +40,7 @@ static int spi_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 }
 
 static int spi_read(comm_context_t *ctx,
-                    uint32_t *rx,
+                    uint8_t *rx,
                     uint16_t rx_sz,
                     uint16_t timeout_ms) {
   (void)ctx;
@@ -54,7 +54,7 @@ static int spi_read(comm_context_t *ctx,
 
 
 static int
-spi_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
+spi_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
   (void)ctx;
   (void)tx;
   (void)timeout_ms;
@@ -64,7 +64,7 @@ spi_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
 }
 
 static int spi_write(comm_context_t *ctx,
-                     uint32_t *tx,
+                     uint8_t *tx,
                      uint16_t tx_size,
                      uint16_t timeout_ms) {
   (void)ctx;
@@ -76,7 +76,7 @@ static int spi_write(comm_context_t *ctx,
   return bytes_written;
 }
 
-static int spi_ioctl(comm_context_t *ctx, uint32_t opcode, void *data) {
+static int spi_ioctl(comm_context_t *ctx, uint8_t opcode, void *data) {
   (void)ctx;
   (void)opcode;
   (void)data;
