+++ dist-fs/comms/uart.c
@@ -15,16 +15,16 @@
 static int uart_init(comm_context_t *ctx);
 static int uart_read_one(comm_context_t *ctx, uint16_t timeout_ms);
 static int uart_read(comm_context_t *ctx,
-                     uint32_t *rx,
+                     uint8_t *rx,
                      uint16_t rx_sz,
                      uint16_t timeout_ms);
 static int
-uart_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms);
+uart_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms);
 static int uart_write(comm_context_t *ctx,
-                      uint32_t *tx,
+                      uint8_t *tx,
                       uint16_t tx_size,
                       uint16_t timeout_ms);
-static int uart_ioctl(comm_context_t *ctx, uint32_t opcode, void *data);
+static int uart_ioctl(comm_context_t *ctx, uint8_t opcode, void *data);
 
 comm_driver_t uart_ops = {
   .init      = uart_init,
@@ -134,7 +134,7 @@ static int uart_read_one(comm_context_t *ctx, uint16_t timeout_ms) {
 }
 
 static int uart_read(comm_context_t *ctx,
-                     uint32_t *rx,
+                     uint8_t *rx,
                      uint16_t rx_sz,
                      uint16_t timeout_ms) {
   if (!ctx || !ctx->driver || !rx)
@@ -168,7 +168,7 @@ static int uart_read(comm_context_t *ctx,
 }
 
 static int
-uart_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
+uart_write_one(comm_context_t *ctx, uint8_t tx, uint16_t timeout_ms) {
   if (!ctx || !ctx->driver)
     return -EINVAL;
 
@@ -185,7 +185,7 @@ uart_write_one(comm_context_t *ctx, uint32_t tx, uint16_t timeout_ms) {
 }
 
 static int uart_write(comm_context_t *ctx,
-                      uint32_t *tx,
+                      uint8_t *tx,
                       uint16_t tx_size,
                       uint16_t timeout_ms) {
   if (!ctx || !ctx->driver || !tx)
@@ -219,7 +219,7 @@ static int uart_write(comm_context_t *ctx,
   }
 }
 
-static int uart_ioctl(comm_context_t *ctx, uint32_t opcode, void *data) {
+static int uart_ioctl(comm_context_t *ctx, uint8_t opcode, void *data) {
   if (!ctx || !ctx->driver)
     return -EINVAL;
